LPD-43072 Inline created Exception
