+++ modules/apps/login/login-web/src/main/resources/META-INF/resources/init.jsp
@@ -45,7 +45,7 @@ page import="com.liferay.portal.kernel.exception.NoSuchUserException" %><%@
 page import="com.liferay.portal.kernel.exception.PasswordExpiredException" %><%@
 page import="com.liferay.portal.kernel.exception.PhoneNumberException" %><%@
 page import="com.liferay.portal.kernel.exception.PhoneNumberExtensionException" %><%@
-page import="com.liferay.portal.kernel.exception.PwdEncryptorAlgorithmException" %><%@
+page import="com.liferay.portal.kernel.exception.PwdEncryptorException" %><%@
 page import="com.liferay.portal.kernel.exception.RequiredFieldException" %><%@
 page import="com.liferay.portal.kernel.exception.RequiredReminderQueryException" %><%@
 page import="com.liferay.portal.kernel.exception.SendPasswordException" %><%@
