+++ modules/apps/login/login-web/src/main/resources/META-INF/resources/login.jsp
@@ -123,7 +123,7 @@
 				<liferay-ui:error exception="<%= CookieNotSupportedException.class %>" message="authentication-failed-please-enable-browser-cookies" />
 				<liferay-ui:error exception="<%= NoSuchUserException.class %>" message="authentication-failed" />
 				<liferay-ui:error exception="<%= PasswordExpiredException.class %>" message="your-password-has-expired" />
-				<liferay-ui:error exception="<%= PwdEncryptorAlgorithmException.class %>" message="authentication-failed" />
+				<liferay-ui:error exception="<%= PwdEncryptorException.class %>" message="authentication-failed" />
 				<liferay-ui:error exception="<%= UserEmailAddressException.MustNotBeNull.class %>" message="please-enter-an-email-address" />
 				<liferay-ui:error exception="<%= UserLockoutException.LDAPLockout.class %>" message="this-account-is-locked" />
