+++ portal-kernel/src/com/liferay/portal/kernel/exception/PwdEncryptorException.java
@@ -25,4 +25,23 @@ public class PwdEncryptorException extends PortalException {
 		super(throwable);
 	}
 
+	public static class PwdEncryptorAlgorithmException extends PwdEncryptorException {
+
+		public PwdEncryptorAlgorithmException() {
+		}
+
+		public PwdEncryptorAlgorithmException(String msg) {
+			super(msg);
+		}
+
+		public PwdEncryptorAlgorithmException(String msg, Throwable throwable) {
+			super(msg, throwable);
+		}
+
+		public PwdEncryptorAlgorithmException(Throwable throwable) {
+			super(throwable);
+		}
+
+	}
+
 }
